Upload images individually due to folder upload limit
Images were uploaded one by one because GitHub does not support uploading entire folders via the web interface.  
The images are originally located in a folder named "image".  
To properly activate image paths in the website, the "image" folder needs to be created in the repository,  
and all images should be placed inside it.